+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-[Unreleased](https://github.com/jshwi/spall/compare/v0.1.1...HEAD)
+[Unreleased](https://github.com/jshwi/spall/compare/v0.2.0...HEAD)
+------------------------------------------------------------------------
+
+[0.2.0](https://github.com/jshwi/spall/releases/tag/v0.2.0) - 2022-05-06
 ------------------------------------------------------------------------
 ### Changed
 - Changes stderr to behave the same way as stdout